Time: 173 ms (9.59%), Space: 8.6 MB (28.75%) - LeetHub

+++ 1358-number-of-substrings-containing-all-three-characters/1358-number-of-substrings-containing-all-three-characters.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     int numberOfSubstrings(string s) {
         int l=0,r=0;
-        long long int n=s.size();
-        unordered_map<char,int> m;
+       int n=s.size();
+        unordered_map<char,long long int> m;
         int res=0;
         while(r<n){
             m[s[r]]++;
